+++ src/pybel/manager/models.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 
 """
+
 This module contains the database models that support the PyBEL definition cache and graph cache
+
 """
 
 import datetime
@@ -10,6 +12,8 @@ from sqlalchemy import Column, ForeignKey, Table, UniqueConstraint
 from sqlalchemy import Integer, String, DateTime, Text, Date, Binary, Boolean
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
+
+from ..constants import CITATION_REFERENCE, CITATION_TYPE, METADATA_VERSION, METADATA_NAME
 
 NAMESPACE_TABLE_NAME = 'pybel_namespace'
 NAMESPACE_ENTRY_TABLE_NAME = 'pybel_namespaceEntry'
@@ -24,13 +28,23 @@ OWL_ANNOTATION_ENTRY_TABLE_NAME = 'pybel_owlAnnotationEntry'
 NAMESPACE_EQUIVALENCE_CLASS_TABLE_NAME = 'pybel_namespaceEquivalenceClass'
 NAMESPACE_EQUIVALENCE_TABLE_NAME = 'pybel_namespaceEquivalence'
 
+CITATION_TABLE_NAME = 'pybel_citation'
+EVIDENCE_TABLE_NAME = 'pybel_evidence'
+NETWORK_EDGE_TABLE_NAME = 'pybel_network_edge'
 NETWORK_TABLE_NAME = 'pybel_network'
+NODE_TABLE_NAME = 'pybel_node'
+EDGE_TABLE_NAME = 'pybel_edge'
+AUTHOR_TABLE_NAME = 'pybel_author'
+AUTHOR_CITATION_TABLE_NAME = 'pybel_author_citation'
+EDGE_ANNOTATION_TABLE_NAME = 'pybel_edge_annotationEntry'
 
 Base = declarative_base()
 
 
 class Namespace(Base):
+    """Represents a BEL Namespace"""
     __tablename__ = NAMESPACE_TABLE_NAME
+
     id = Column(Integer, primary_key=True)
 
     url = Column(String(255))
@@ -59,7 +73,9 @@ class Namespace(Base):
 
 
 class NamespaceEntry(Base):
+    """Represents a name within a BEL namespace"""
     __tablename__ = NAMESPACE_ENTRY_TABLE_NAME
+
     id = Column(Integer, primary_key=True)
 
     name = Column(String(255), nullable=False)
@@ -74,6 +90,7 @@ class NamespaceEntry(Base):
 
 class NamespaceEntryEquivalence(Base):
     __tablename__ = NAMESPACE_EQUIVALENCE_CLASS_TABLE_NAME
+
     id = Column(Integer, primary_key=True)
     label = Column(String(255), nullable=False, unique=True, index=True)
 
@@ -81,6 +98,7 @@ class NamespaceEntryEquivalence(Base):
 
 
 class Annotation(Base):
+    """Represents a BEL Annotation"""
     __tablename__ = ANNOTATION_TABLE_NAME
 
     id = Column(Integer, primary_key=True)
@@ -108,7 +126,9 @@ class Annotation(Base):
 
 
 class AnnotationEntry(Base):
+    """Represents a value within a BEL Annotation"""
     __tablename__ = ANNOTATION_ENTRY_TABLE_NAME
+
     id = Column(Integer, primary_key=True)
 
     name = Column(String(255), nullable=False)
@@ -126,6 +146,7 @@ owl_namespace_relationship = Table(
 
 
 class OwlNamespace(Base):
+    """Represents an OWL Namespace"""
     __tablename__ = OWL_NAMESPACE_TABLE_NAME
 
     id = Column(Integer, primary_key=True)
@@ -135,6 +156,7 @@ class OwlNamespace(Base):
 
 
 class OwlNamespaceEntry(Base):
+    """Represents a name within an OWL Namespace"""
     __tablename__ = OWL_NAMESPACE_ENTRY_TABLE_NAME
 
     id = Column(Integer, primary_key=True)
@@ -159,6 +181,7 @@ owl_annotation_relationship = Table(
 
 
 class OwlAnnotation(Base):
+    """Represents an OWL namespace used as an annotation"""
     __tablename__ = OWL_ANNOTATION_TABLE_NAME
 
     id = Column(Integer, primary_key=True)
@@ -168,6 +191,7 @@ class OwlAnnotation(Base):
 
 
 class OwlAnnotationEntry(Base):
+    """Represents a name in an OWL namespace used as an annotation"""
     __tablename__ = OWL_ANNOTATION_ENTRY_TABLE_NAME
 
     id = Column(Integer, primary_key=True)
@@ -184,8 +208,29 @@ class OwlAnnotationEntry(Base):
                             secondaryjoin=id == owl_annotation_relationship.c.right_id)
 
 
+network_edge = Table(
+    NETWORK_EDGE_TABLE_NAME, Base.metadata,
+    Column('network_id', Integer, ForeignKey('{}.id'.format(NETWORK_TABLE_NAME)), primary_key=True),
+    Column('edge_id', Integer, ForeignKey('{}.id'.format(EDGE_TABLE_NAME)), primary_key=True)
+)
+
+author_citation = Table(
+    AUTHOR_CITATION_TABLE_NAME, Base.metadata,
+    Column('author_id', Integer, ForeignKey('{}.id'.format(AUTHOR_TABLE_NAME))),
+    Column('citation_id', Integer, ForeignKey('{}.id'.format(CITATION_TABLE_NAME)))
+)
+
+edge_annotation = Table(
+    EDGE_ANNOTATION_TABLE_NAME, Base.metadata,
+    Column('edge_id', Integer, ForeignKey('{}.id'.format(EDGE_TABLE_NAME)), primary_key=True),
+    Column('annotationEntry_id', Integer, ForeignKey('{}.id'.format(ANNOTATION_ENTRY_TABLE_NAME)), primary_key=True)
+)
+
+
 class Network(Base):
+    """Represents a collection of edges, specified by a BEL Script"""
     __tablename__ = NETWORK_TABLE_NAME
+
     id = Column(Integer, primary_key=True)
 
     name = Column(String(255), index=True)
@@ -201,6 +246,78 @@ class Network(Base):
     created = Column(DateTime, default=datetime.datetime.utcnow)
     blob = Column(Binary)
 
+    edges = relationship('Edge', secondary=network_edge)
+
     __table_args__ = (
-        UniqueConstraint("name", "version"),
+        UniqueConstraint(METADATA_NAME, METADATA_VERSION),
     )
+
+
+class Node(Base):
+    """Represents a BEL Term"""
+    __tablename__ = NODE_TABLE_NAME
+
+    id = Column(Integer, primary_key=True)
+
+    bel = Column(String, nullable=False)
+    blob = Column(Binary)
+
+
+class Edge(Base):
+    """Represents the relation between two BEL terms and its properties"""
+    __tablename__ = EDGE_TABLE_NAME
+
+    id = Column(Integer, primary_key=True)
+
+    source_id = Column(Integer, ForeignKey('{}.id'.format(NODE_TABLE_NAME)))
+    source = relationship('Node', foreign_keys=[source_id])
+
+    target_id = Column(Integer, ForeignKey('{}.id'.format(NODE_TABLE_NAME)))
+    target = relationship('Node', foreign_keys=[target_id])
+
+    evidence_id = Column(Integer, ForeignKey('{}.id'.format(EVIDENCE_TABLE_NAME)))
+    evidence = relationship("Evidence")
+
+    annotations = relationship('AnnotationEntry', secondary=edge_annotation)
+
+    relation = Column(String, nullable=False)
+    bel = Column(String, nullable=False)
+    blob = Column(Binary)
+
+
+class Evidence(Base):
+    """Represents a piece of support taken from a Publication"""
+    __tablename__ = EVIDENCE_TABLE_NAME
+    id = Column(Integer, primary_key=True)
+    text = Column(String, nullable=False, index=True)
+
+    citation_id = Column(Integer, ForeignKey('{}.id'.format(CITATION_TABLE_NAME)))
+    citation = relationship('Citation')
+
+
+class Citation(Base):
+    """The information about the citations that are used to prove a specific relation are stored in this table."""
+    __tablename__ = CITATION_TABLE_NAME
+
+    id = Column(Integer, primary_key=True)
+    type = Column(String(16), nullable=False)
+    name = Column(String(255), nullable=False)
+    reference = Column(String(255), nullable=False)
+    date = Column(Date, nullable=True)
+    comments = Column(String(255), nullable=True)
+
+    authors = relationship("Author", secondary=author_citation)
+
+    __table_args__ = (
+        UniqueConstraint(CITATION_TYPE, CITATION_REFERENCE),
+    )
+
+
+class Author(Base):
+    """Represents an Author of a publication"""
+    __tablename__ = AUTHOR_TABLE_NAME
+
+    id = Column(Integer, primary_key=True)
+    name = Column(String(255), nullable=False)
+
+    citations = relationship("Citation", secondary=author_citation)