+++ src/pybel/parser/parse_control.py
@@ -25,12 +25,18 @@ log = logging.getLogger('pybel')
 
 
 class ControlParser(BaseParser):
-    def __init__(self, annotation_dicts=None, annotation_expressions=None):
+    def __init__(self, annotation_dicts=None, annotation_expressions=None, citation_clearing=True):
         """Builds parser for BEL valid_annotations statements
 
-        :param annotation_dicts: A dictionary from {annotation: set of valid values} for parsing
+        :param annotation_dicts: A dictionary of {annotation: set of valid values} for parsing
         :type annotation_dicts: dict
+        :param annotation_expressions: A dictionary of {annotation: regular expression string}
+        :type annotation_expressions: dict
+        :param citation_clearing: Should :code:`SET Citation` statements clear evidence and all annotations?
+        :type citation_clearing: bool
         """
+
+        self.citation_clearing = citation_clearing
 
         self.valid_annotations = {} if annotation_dicts is None else annotation_dicts
         self.annotations_re = {} if annotation_expressions is None else annotation_expressions
@@ -102,10 +108,6 @@ class ControlParser(BaseParser):
 
         BaseParser.__init__(self, self.language)
 
-    def raise_if_no_citation(self, s):
-        if not self.citation:
-            raise MissingCitationException(s)
-
     def validate_annotation_key(self, key):
         if key not in self.valid_annotations and key not in self.annotations_re_compiled:
             raise UndefinedAnnotationWarning(key)
@@ -119,7 +121,10 @@ class ControlParser(BaseParser):
     def handle_annotation_key(self, s, l, tokens):
         """Called on all annotation keys before parsing to validate that it's either enumerated or as a regex"""
         key = tokens['key']
-        self.raise_if_no_citation(s)
+
+        if self.citation_clearing and not self.citation:
+            raise MissingCitationException(s)
+
         self.validate_annotation_key(key)
         return tokens
 
@@ -177,7 +182,9 @@ class ControlParser(BaseParser):
     def handle_unset_citation(self, s, l, tokens):
         if not self.citation:
             raise MissingAnnotationKeyWarning(BEL_KEYWORD_CITATION)
-        self.citation.clear()
+
+        self.clear_citation()
+
         return tokens
 
     def handle_unset_evidence(self, s, l, tokens):
@@ -224,10 +231,14 @@ class ControlParser(BaseParser):
 
     def clear_citation(self):
         self.citation.clear()
-        self.evidence = None
-        self.annotations.clear()
+
+        if self.citation_clearing:
+            self.evidence = None
+            self.annotations.clear()
 
     def clear(self):
-        """Clears the annotations, citation, and statement group"""
+        """Clears the statement_group, citation, evidence, and annotations"""
         self.statement_group = None
-        self.clear_citation()
+        self.citation.clear()
+        self.evidence = None
+        self.annotations.clear()